playing with cross entropy experiment

+++ experiments/cross_entropy_experiments.py
@@ -5,7 +5,7 @@ from torch.nn.functional import cross_entropy
 def cross_entropy_loss(logits, targets):
   return cross_entropy(logits, targets, reduction='mean')
 
-target = torch.randint(0,5, (1,), dtype=torch.long)
+target = torch.randint(0,5, (1,), dtype=torch.int32)
 
 print(target.shape)
 
@@ -13,5 +13,5 @@ logits = torch.randn(1, 5)
 
 print(logits.shape)
 
-loss = cross_entropy_loss(logits, target)
+loss = cross_entropy_loss(logits, target.long())
 print(loss)
